Rename "dom" fork to "dom-node" and "bun" fork to "dom-bun" (#26117)
The "dom" configuration is actually the node specific configuration. It
just happened to be that this was the mainline variant before so it was
implied but with so many variants, this is less obvious now.

The "bun" configuration is specifically for "bun". There's no "native"
renderer for "bun" yet.

DiffTrain build for [03a216070dfe14ff8fecaee2fa95ca6d6e94846d](https://github.com/facebook/react/commit/03a216070dfe14ff8fecaee2fa95ca6d6e94846d)
[View git log for this commit](https://github.com/facebook/react/commits/03a216070dfe14ff8fecaee2fa95ca6d6e94846d)